+++ lib/index.d.ts
@@ -8,4 +8,7 @@ export { Menu, type MenuProps, type MenuItemProps } from './Menu';
 export { Slider, type SliderProps, type SliderOption } from './Slider';
 export { Tabs, type TabsProps, type TabItemProps } from './Tabs';
 export { TextField, type TextFieldProps } from './TextField';
+export { Checkbox } from './Checkbox';
+export { Switch } from './Switch';
+export { Radio } from './Radio';
 //# sourceMappingURL=index.d.ts.map