+++ vite.config.ts
@@ -7,7 +7,7 @@ import { name, version } from './package.json';
 // https://vitejs.dev/config/
 export default defineConfig({
   server: {
-    port: 8000,
+    port: 4000,
   },
   build: {
     outDir: 'docs/theme',